+++ api/src/team/domain/usecases/index.js
@@ -2,6 +2,7 @@ import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import * as mailService from '../../../../lib/domain/services/mail-service.js';
+import * as organizationRepository from '../../../../src/shared/infrastructure/repositories/organization-repository.js';
 import * as certificationCenterRepository from '../../../certification/shared/infrastructure/repositories/certification-center-repository.js';
 import * as membershipRepository from '../../../shared/infrastructure/repositories/membership-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
@@ -9,7 +10,9 @@ import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/
 import * as certificationCenterInvitationRepository from '../../infrastructure/repositories/certification-center-invitation-repository.js';
 import { prescriberRepository } from '../../infrastructure/repositories/prescriber-repository.js';
 import { userOrgaSettingsRepository } from '../../infrastructure/repositories/user-orga-settings-repository.js';
+import { organizationInvitationRepository } from '../../infrastructure/repositories/organization-invitation.repository.js';
 import * as certificationCenterInvitationService from '../services/certification-center-invitation-service.js';
+import { organizationInvitationService } from '../services/organization-invitation.service.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
 
@@ -21,6 +24,9 @@ const dependencies = {
   userOrgaSettingsRepository,
   certificationCenterInvitationService,
   mailService,
+  organizationInvitationService,
+  organizationInvitationRepository,
+  organizationRepository,
 };
 
 const usecasesWithoutInjectedDependencies = {
