+++ api/tests/integration/application/organization-invitations/organization-invitation-controller_test.js
@@ -1,13 +1,5 @@
 import * as moduleUnderTest from '../../../../lib/application/organization-invitations/index.js';
-import {
-  AlreadyExistingInvitationError,
-  ManyOrganizationsFoundError,
-  NotFoundError,
-  OrganizationArchivedError,
-  OrganizationNotFoundError,
-  OrganizationWithoutEmailError,
-  UserNotFoundError,
-} from '../../../../lib/domain/errors.js';
+import { AlreadyExistingInvitationError, NotFoundError, UserNotFoundError } from '../../../../lib/domain/errors.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
 import { serializer as scoOrganizationInvitationSerializer } from '../../../../lib/infrastructure/serializers/jsonapi/sco-organization-invitation-serializer.js';
 import { expect, HttpTestServer, sinon } from '../../../test-helper.js';
@@ -19,7 +11,6 @@ describe('Integration | Application | Organization-invitations | organization-in
   beforeEach(async function () {
     sandbox = sinon.createSandbox();
     sandbox.stub(usecases, 'acceptOrganizationInvitation');
-    sandbox.stub(usecases, 'sendScoInvitation');
     sandbox.stub(usecases, 'createCertificationCenterMembershipForScoOrganizationAdminMember');
     sandbox.stub(scoOrganizationInvitationSerializer, 'serialize');
 
@@ -89,79 +80,6 @@ describe('Integration | Application | Organization-invitations | organization-in
 
         // then
         expect(response.statusCode).to.equal(404);
-      });
-    });
-  });
-
-  describe('#sendScoInvitation', function () {
-    const uai = '1234567A';
-    const payload = {
-      data: {
-        type: 'sco-organization-invitations',
-        attributes: {
-          uai,
-          'first-name': 'john',
-          'last-name': 'harry',
-        },
-      },
-    };
-
-    context('Success cases', function () {
-      it('should return an HTTP response with status code 201', async function () {
-        // given
-        usecases.sendScoInvitation.resolves();
-
-        // when
-        const response = await httpTestServer.request('POST', '/api/organization-invitations/sco', payload);
-
-        // then
-        expect(response.statusCode).to.equal(201);
-      });
-    });
-
-    context('Error cases', function () {
-      it('should respond an HTTP response with status code 404 when OrganizationNotFoundError', async function () {
-        // given
-        usecases.sendScoInvitation.rejects(new OrganizationNotFoundError());
-
-        // when
-        const response = await httpTestServer.request('POST', '/api/organization-invitations/sco', payload);
-
-        // then
-        expect(response.statusCode).to.equal(404);
-      });
-
-      it('should respond an HTTP response with status code 412 when OrganizationWithoutEmailError', async function () {
-        // given
-        usecases.sendScoInvitation.rejects(new OrganizationWithoutEmailError());
-
-        // when
-        const response = await httpTestServer.request('POST', '/api/organization-invitations/sco', payload);
-
-        // then
-        expect(response.statusCode).to.equal(412);
-      });
-
-      it('should respond an HTTP response with status code 409 when ManyOrganizationsFoundError', async function () {
-        // given
-        usecases.sendScoInvitation.rejects(new ManyOrganizationsFoundError());
-
-        // when
-        const response = await httpTestServer.request('POST', '/api/organization-invitations/sco', payload);
-
-        // then
-        expect(response.statusCode).to.equal(409);
-      });
-
-      it('should respond an HTTP response with status code 422 when OrganizationArchivedError', async function () {
-        // given
-        usecases.sendScoInvitation.rejects(new OrganizationArchivedError());
-
-        // when
-        const response = await httpTestServer.request('POST', '/api/organization-invitations/sco', payload);
-
-        // then
-        expect(response.statusCode).to.equal(422);
       });
     });
   });
